Add example API call and show example screenshot
Added an example API call which the client devices can use for verifying a package drop off. In the actual application, the QR scanner will read the  user's primary key. It will use that PK to call the API to check if it is the right box for the delivery.